feat: BaseListView 구현 및 상속
#119
